fix type separation
