+++ Package.swift
@@ -13,6 +13,7 @@ let package = Package(
       name: "ExtractCaseValue",
       targets: ["ExtractCaseValue"]
     ),
+    .library(name: "ExtractCaseValueTypes", targets: ["ExtractCaseValueTypes"]),
     .executable(
       name: "ExtractCaseValueClient",
       targets: ["ExtractCaseValueClient"]
@@ -30,12 +31,16 @@ let package = Package(
       name: "ExtractCaseValueMacros",
       dependencies: [
         .product(name: "SwiftSyntaxMacros", package: "swift-syntax"),
-        .product(name: "SwiftCompilerPlugin", package: "swift-syntax")
+        .product(name: "SwiftCompilerPlugin", package: "swift-syntax"),
+        "ExtractCaseValueTypes"
       ]
     ),
 
     // Library that exposes a macro as part of its API, which is used in client programs.
-    .target(name: "ExtractCaseValue", dependencies: ["ExtractCaseValueMacros"]),
+    .target(name: "ExtractCaseValue", dependencies: ["ExtractCaseValueMacros", "ExtractCaseValueTypes"]),
+    .target(name: "ExtractCaseValueTypes", dependencies: [
+        .product(name: "SwiftSyntax", package: "swift-syntax"),
+    ]),
 
     // A client of the library, which is able to use the macro in its own code.
     .executableTarget(name: "ExtractCaseValueClient", dependencies: ["ExtractCaseValue"]),
