+++ Sources/ExtractCaseValue/ExtractCaseValue.swift
@@ -1,4 +1,6 @@
-import ExtractCaseValueMacros
+
+//import ExtractCaseValueMacros
+import ExtractCaseValueTypes
 
 /// A macro that extracts an associated value from enum cases using a default value if
 /// extraction is not possible.
